requirements.txt
certstream==1.11
dnspython==2.0.0
six==1.15.0
termcolor==1.1.0
websocket-client==0.57.0
